# AccSol.React/Data/swagger2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Accsol.API",
    "version": "v1"
  },
  "paths": {
...
    "/ProjectCategories/Get": {
      "post": {
        "tags": [
          "ProjectCategories"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCategory"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCategory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCategory"
                }
              }
            }
          }
        }
      }
    },
    "/ProjectCategories/Save": {
      "post": {
        "tags": [
          "ProjectCategories"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCategory"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCategory"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCategory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ProjectCategories/Delete": {
      "post": {
        "tags": [
          "ProjectCategories"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/ProjectCodes/Get": {
      "post": {
        "tags": [
          "ProjectCodes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCode"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCode"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCode"
                }
              }
            }
          }
        }
      }
    },
    "/ProjectCodes/Save": {
      "post": {
        "tags": [
          "ProjectCodes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCode"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCode"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCode"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ProjectCodes/Delete": {
      "post": {
        "tags": [
          "ProjectCodes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProjectCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProjectCode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "projectCategoryID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "projectYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "clientID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "projectNumber": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}